+++ rpm_lockfile/utils.py
@@ -1,3 +1,4 @@
+import json
 import logging
 import os
 import shlex
@@ -55,3 +56,37 @@ def get_file_from_git(repo, ref, file):
         # logged.
         subprocess.run(cmd, cwd=tmp_dir, check=True)
     return os.path.join(tmp_dir, file)
+
+
+def subst_vars(template, vars):
+    """Replace {var} placeholders in template with provided values."""
+    for key, value in vars.items():
+        template = template.replace(f"{{{key}}}", value)
+    return template
+
+
+def _get_image_labels(image_spec):
+    """Given an image specification, return a dict with labels from the image."""
+    if not image_spec:
+        return {}
+    cp = logged_run(
+        ["skopeo", "inspect", f"docker://{image_spec}"],
+        stdout=subprocess.PIPE,
+        check=True,
+    )
+    data = json.loads(cp.stdout)
+    return data["Labels"]
+
+
+def _get_containerfile_labels(containerfile):
+    """Find labels of the last base image used in the given containerfile."""
+    if not containerfile:
+        return {}
+    if not containerfile.startswith("/"):
+        raise ValueError("Containerfile must be specified by absolute path")
+    return _get_image_labels(extract_image(containerfile))
+
+
+def get_labels(image_spec, containerfile):
+    """Find labels from given image or the base image used in the containerfile."""
+    return _get_image_labels(image_spec) | _get_containerfile_labels(containerfile)